+++ R/getPsiteCoordinates.R
@@ -13,7 +13,7 @@
 #' @export
 #' @examples
 #' library(Rsamtools)
-#' bamfilename <- system.file("extdata", "RPF.chr1.bam",
+#' bamfilename <- system.file("extdata", "RPF.WT.1.bam",
 #'                            package="ribosomeProfilingQC")
 #' yieldSize <- 10000000
 #' bamfile <- BamFile(bamfilename, yieldSize = yieldSize)
@@ -40,6 +40,7 @@ getPsiteCoordinates <- function(bamfile, bestpsite, anchor='5end'){
   reads <- as(reads, "GRanges")
   reads <- promoters(reads, upstream = 0, downstream = 1)
   reads$Psite <- bestpsite
+  if(anchor=="3end") reads$Psite <- reads$qwidth - reads$Psite
   metadata(reads) <- list(totalReads=length(reads))
   reads
 }