+++ R/prepareCDS.R
@@ -18,9 +18,14 @@
 #' @import GenomicRanges
 #' @export
 #' @examples
-#' library(AnnotationDbi)
-#' txdb <- loadDb(system.file("extdata", "danRer10.chr1.txdb",
-#'                package="ribosomeProfilingQC"))
+#' library(GenomicFeatures)
+#' library(BSgenome.Drerio.UCSC.danRer10)
+#' txdb <- makeTxDbFromGFF(system.file("extdata",
+#'           "Danio_rerio.GRCz10.91.chr1.gtf.gz",
+#'           package="ribosomeProfilingQC"),
+#'           organism = "Danio rerio",
+#'           chrominfo = seqinfo(Drerio)["chr1"],
+#'           taxonomyId = 7955)
 #' CDS <- prepareCDS(txdb)
 #'
 prepareCDS <- function(txdb, withUTR = FALSE){