+++ NEWS.md
@@ -1,3 +1,11 @@
+# CHANGES IN VERSION 0.99.1
+
+* decrease the size of the package
+* simplify and add parameter anchor for the estimatePsite function.
+* add parameter normByLibSize for function translationalEfficiency
+* add more test files.
+* update documentations.
+
 # CHANGES IN VERSION 0.99.0
 
 * Submit to Bioconductor.